chat/requirements.txt
asgiref==3.3.1
certifi==2020.12.5
chardet==4.0.0
Django==3.1.5
django-cors-headers==3.6.0
django-environ==0.4.5
djangorestframework==3.12.2
djangorestframework-simplejwt==4.6.0
gunicorn==20.0.4
idna==2.10
psycopg2-binary==2.8.6
PyJWT==2.0.1
pytz==2020.5
requests==2.25.1
sqlparse==0.4.1
urllib3==1.26.2
